initrd

+++ kernel/filesystem/initrd.h
@@ -3,14 +3,14 @@
 #include "./file.h"
 
 typedef struct {
-	uint32_t nfiles; // The number of files in the ramdisk.
+	size_t nfiles; // The number of files in the ramdisk.
 } initrd_header_t;
 
 typedef struct {
 	uint8_t magic;	    // Magic number, for error checking.
 	char name[64];      // Filename.
-	uint32_t offset;	// Offset in the initrd that the file starts.
-	uint32_t length;	// Length of the file.
+	size_t offset;	// Offset in the initrd that the file starts.
+	size_t length;	// Length of the file.
 } initrd_file_header_t;
 
 // Initialises the initial ramdisk. It gets passed the address of the multiboot module,
